backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
sqlalchemy==2.0.40
pymysql==1.1.1
python-dotenv==1.0.1
